+++ NEWS.md
@@ -6,6 +6,7 @@ like a more useful default.
 a string format can be executed conditionally.
 * Added "type" parameter to `fmt_quantile_range()` function so user can
 specify the type of quantile algorithm to use.
+* Updated logo.
 
 # fmtr 1.5.8